requirements.txt
streamlit==1.35.0
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.2
langchain==0.2.0
openai==1.30.1
llama-index==0.10.20
pandas==2.2.2
python-dotenv==1.0.1
google-generativeai==0.3.2
httpx==0.27.0
